Move set_log_level to Xav module; rename config key to :ffmpeg_log_level
